requirements.txt
fastapi~=0.78.0
minio==7.1.3
psycopg2_binary==2.9.6
uvicorn~=0.18.2
stripe==3.4.0 
numpy-stl
python-multipart
keycloak~=3.1.3
pydantic~=1.9.1
requests~=2.28.1
selenium~=4.3.0
numpy~=1.23.0
markdownify~=0.11.2
easypost~=7.4.0
python-keycloak
boto3
filetype
